+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 | Version      |    Date     | Notes |
 |--------------|:-----------:|-----------------------------------------------------|
+| 3.0.2        |   2/29/24   | Updated evaluation library to version 2.0.3
 | 3.0.1        |   2/7/24    | Updated evaluation library
 | 3.0          |   1/25/24   | Fixed bugs and add additional WCGA 2.1 and 2.2 rules
 | 2.99.6       |   12/15/23  | Fixed bugs and add target size rules
@@ -19,6 +20,10 @@
 | 1.1.2        |   9/20/21   | Updated WIDGET 13 documentation and included secondary WCAG SC in rule details
 | 1.1.1        |   9/16/21   | Updated the extension to support ARIA 1.2 and ARIA in HTML requirements
 | 0.96.0       |  10/30/19   | Fixes bug in highlight option not being remembered
+
+=======
+### 3.02
+* Update includes fixed bug in color contrast calculation
 
 ### 3.01
 * Updated evaluation library to versino 2.0.2 with improved support of row role in table, grid and treegrid; and color contrast rule